JWT + move files to Utils folder
