+++ TODO.md
@@ -1,5 +1,6 @@
 # 0.0.1
 
-- [ ] Hacer el login
+- [x] Hacer el login
+- [ ] JWT
 - [ ] Encriptar el usuario y contraseña
 - [ ] Probar el despliegue y fixearlo ya que creo que con los nuevo script no funcionara bien
