+++ package.json
@@ -35,10 +35,11 @@
     "typescript": "^3.7.5"
   },
   "dependencies": {
+    "@prisma/cli": "^2.3.0",
     "body-parser": "^1.19.0",
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
-    "regenerator-runtime": "^0.13.7",
-    "@prisma/cli": "^2.3.0"
+    "jsonwebtoken": "^8.5.1",
+    "regenerator-runtime": "^0.13.7"
   }
 }
