+++ src/Utils/env.constants.ts
@@ -1,4 +1,5 @@
 export const envConstants = {
   NODE_ENV: process.env.NODE_ENV,
   PORT: 5000,
+  JWT: 'test',
 };
